+++ Changelog.md
@@ -1,4 +1,9 @@
 
+# UNRELEASED
+
+### HOAS:
+- New `cofix` and `mcofix`
+
 # [3.5.1] 04/09/2026
 
 Requires Elpi 3.7.1 and Rocq 9.0, 9.1, 9.2 and 9.3.